requirements.txt
groq
streamlit
requests
Pillow
python-dotenv
